+++ app/Http/Controllers/Cursosinternos/LeccionesController.php
@@ -3,28 +3,37 @@
 namespace App\Http\Controllers\Cursosinternos;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\SaveLeccionRequest;
 use App\Models\Leccion;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
 class LeccionesController extends Controller
 {
-    
+
     public function destroy(string $id)
     {
         $leccione = Leccion::find($id);
-        $leccione->delete();
-        return redirect()->back();
+        $id_contenido = $leccione->contenido->pluck('id_contenido')->toArray();
+        $coincidencias = DB::table('contenidos')
+        ->whereIn('id_contenido',$id_contenido)
+        ->exists();
+        if (!$coincidencias) {
+            $leccione->delete();
+            return redirect()->back()->with('eliminado', 'Eliminado Correctamente');
+        }
+        return redirect()->back()->with('error', 'No se puede eliminar el registro porque está asociado a otro campo.');
     }
 
     public function show(string $id)
     {
-        return view('Cursosinternos.lecciones.agregar',compact('id'));
+        return view('Cursosinternos.lecciones.agregar', compact('id'));
     }
     public function edit(string $id)
     {
         $leccion = Leccion::find($id);
-        return view('Cursosinternos.lecciones.editar',compact('leccion'));
+        return view('Cursosinternos.lecciones.editar', compact('leccion'));
     }
 
     public function update(Request $request, string $id)
@@ -40,11 +49,10 @@ class LeccionesController extends Controller
         $leccione->curso_id = $request->post('curso_id');
         $leccione->saveOrFail();
         $curso = $leccione->curso_id;
-        return to_route("curs.show",$curso)->with('agregado', 'Leccion Actualizada Correctamente');
-
+        return to_route("curs.show", $curso)->with('agregado', 'Leccion Actualizada Correctamente');
     }
 
-    public function store(Request $request)
+    public function store(SaveLeccionRequest $request)
     {
         $leccione = new Leccion();
         $leccione->nombre = $request->post('nombre');
@@ -55,6 +63,6 @@ class LeccionesController extends Controller
         $leccione->url_imagen = $url;
         $leccione->saveOrFail();
         $curso = $leccione->curso_id;
-        return to_route("curs.show",$curso)->with('agregado', 'Leccion Agregado Correctamente');
+        return to_route("curs.show", $curso)->with('agregado', 'Leccion Agregado Correctamente');
     }
 }
